+++ CHANGELOG.md
@@ -5,6 +5,19 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [3.53.2] - 2023-02-22
+
+### Fixed
+
+- rendering of empty message
+
+## [3.53.0] - 2023-02-21
+
+### Added
+
+- variable expires with a new session `vars.sessionContext()`
+- sequence messages
+
 ## [3.52.6] - 2023-02-18
 
 ### Fixed